+++ src/middleware/mod.rs
@@ -5,15 +5,15 @@ mod compress;
 #[cfg(feature = "compress")]
 pub use self::compress::Compress;
 
+mod compat;
 mod condition;
 mod defaultheaders;
 pub mod errhandlers;
 mod logger;
 pub mod normalize;
-mod scoped;
 
+pub use self::compat::Compat;
 pub use self::condition::Condition;
 pub use self::defaultheaders::DefaultHeaders;
 pub use self::logger::Logger;
 pub use self::normalize::NormalizePath;
-pub use self::scoped::Scoped;